refactor: search path test as a integration test
